:left_right_arrow: split localstorage tests into own file
